refactor(backend): update the standalone scripts to the new module paths
scripts/{migrate_starlib_meta,build_bpm_fixture,generate_sample_tracks}.py
still imported backend.core.* and backend.sc_settings.
